+++ premium/web-frontend/modules/baserow_premium/components/admin/user/UserAdminTable.vue
@@ -35,7 +35,7 @@ import LocalDateField from '@baserow_premium/components/crud_table/fields/LocalD
 import ActiveField from '@baserow_premium/components/admin/user/fields/ActiveField'
 import EditUserContext from '@baserow_premium/components/admin/user/contexts/EditUserContext'
 import HiddenGroupsContext from '@baserow_premium/components/admin/user/contexts/HiddenGroupsContext'
-import CrudTableColumn from '@baserow_premium/components/crud_table/CrudTableColumn'
+import CrudTableColumn from '@baserow_premium/crud_table/CrudTableColumn'
 
 export default {
   name: 'UserAdminTable',