+++ changelog.md
@@ -13,9 +13,7 @@
 * Make the view header more compact when the content doesn't fit anymore.
 * Allow providing a `template_id` when registering a new account, which will install
   that template instead of the default database.
-
-## Unreleased Premium Only Features
-* Added user admin area allowing management of baserow users.
+* Premium Only: Added user admin area allowing management of baserow users.
 
 ## Released (2021-04-08)